Add Mock views, DB config, model & style fixes
Add CRUD Razor views for Mock/Cliente (Create, Delete, Details, Edit, Index) and integrate _TabelaAcoes partial in the index. Change Veiculo.PrecoDia from Decimal to float. Extract connection string into a variable in Program.cs and add DefaultConnection to appsettings.json for localdb. Update Shared/_Layout to point the Clientes nav item to the Mock controller. Tweak dark theme CSS (color-scheme, themeSelector border and focus styles).

+++ Data/Projeto_AutoMobile/Veiculo.cs
@@ -24,7 +24,7 @@ namespace Projeto_AutoMobile.Data.Projeto_AutoMobile
         public string Matricula { get; set; } = string.Empty;
         public string Marca {  get; set; } = string.Empty;
         public string Modelo { get; set; } = string.Empty;
-        public Decimal PrecoDia { get; set; }
+        public float PrecoDia { get; set; }
         public EstadoVeiculo Estado { get; set; } //o Estado é do tipo EtsadoVeiculo
 
         //data prevista para passar ao estado de “disponível”
